refactor: rename downsampled_views.py → mosaic_utils.py and drop dead helpers
Final cleanup pass on the leftovers from the old plate-view pipeline:

- control/core/downsampled_views.py → control/core/mosaic_utils.py. Module
  docstring rewritten; the old name was misleading after the unified-mosaic
  refactor (no DownsampledView* anything in there).
- Deleted four functions that no production code calls anymore:
    crop_overlap, downsample_to_resolutions, stitch_tiles,
    ensure_plate_resolution_in_well_resolutions
  (~250 LOC removed). Their tests in test_mosaic_utils.py go too — 19 tests.
  Survivors (calculate_overlap_pixels, _pyrdown_chain, downsample_tile,
  parse_well_id, format_well_id) keep the same signatures and tests.
- Pruned the now-unused imports (os, tifffile, ZProjectionMode, Dict/Optional/
  Union typing aliases) from mosaic_utils.py.
- Updated import paths everywhere: gui_hcs.py, widgets_mosaic.py,
  multi_point_worker.py, widgets.py, tests/control/core/test_mosaic_utils.py.
- Renamed the matching key in acquisition.yaml's
  multi_point_controller.dump from "downsampled_views" to "mosaic_view"
  for consistency with the widget's terminology.

Docs:
- Removed docs/downsampled-plate-view.md (entire doc described the deleted
  DISPLAY_PLATE_VIEW flag, multi-resolution wells/A1_5um.tiff outputs, etc.).
- Removed the cross-reference from docs/zarr-v3-format.md.

Tests:
- test_def.py:108-113 — reworded a docstring example that name-dropped
  ensure_plate_resolution_in_well_resolutions and DOWNSAMPLED_WELL_RESOLUTIONS_UM
  in a comment-stripping regression test (cosmetic).

Full suite: 1106 passed, 5 skipped, 1 xfailed. Smoke test clean.
